% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cum_pi_matrix_cpp}
\alias{cum_pi_matrix_cpp}
\title{Make cumulative pi matrix.}
\usage{
cum_pi_matrix_cpp(pi)
}
\arguments{
\item{pi}{The matrix of probabilities}
}
\value{
The cumulative distribution matrix.
}
\description{
Make cumulative pi matrix.
}
