% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cum_pi_to_pi}
\alias{cum_pi_to_pi}
\title{Transform cumulative pi matrix to pi matrix.}
\usage{
cum_pi_to_pi(cumpi)
}
\arguments{
\item{cumpi}{Cumulative pi matrix.}
}
\value{
pi matrix.
}
\description{
Transform cumulative pi matrix to pi matrix.
}
