% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{generate_pi}
\alias{generate_pi}
\title{Generate pi matrices.}
\usage{
generate_pi(i, j, cutoffs = c("uniform"), copula)
}
\arguments{
\item{i, j}{The number of rows and number of columns of the pi matrix.}

\item{cutoffs}{Method to generate the uniform cutoffs. Only \code{"uniform"} right
now.}

\item{copula}{The copula to generate probabilities from.}
}
\value{
The pi matrix.
}
\description{
Generate pi matrices.
}
\examples{
  norm_cop = copula::normalCopula(0.3, dim = 2)
  copula = function(u) copula::pCopula(u, norm_cop)
  pi = polyiden::generate_pi(i = 10, j = 4, copula = copula)
  polycor::polychor(pi) # [1] 0.2999905
}
