% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{get_interior}
\alias{get_interior}
\title{Find interior point.}
\usage{
get_interior(ui, ci, candidate, max_rep)
}
\arguments{
\item{ui}{Constraints matrix.}

\item{ci}{Bounds vector.}

\item{candidate}{Candidate solution vector.}

\item{max_rep}{Maximal times to randomly search for new solution.}
}
\description{
Find interior point.
}
