% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyiden.R
\name{latent_correlation}
\alias{latent_correlation}
\title{Parametric latent correlations}
\usage{
latent_correlation(
  pi,
  distribution = c("normal"),
  discrepancy = function(x, y) -y * log(x)
)
}
\arguments{
\item{pi}{Matrix of probabilities.}

\item{distribution}{The joint distribution of the latent variables,
parameterized by the correlation `"rho"`.}

\item{discrepancy}{Discrepancy function to minimize. Defaults to
`-y * log(x))`, which corresponds to minimization of the Kullback-Leibler
divergence.}
}
\value{
The closest latent correlation as judged by the discrepancy function.
}
\description{
Parametric latent correlations
}
