% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polychoric_pi_set.R
\name{polychoric_pi_set}
\alias{polychoric_pi_set}
\title{Calculate the partial identification set for the latent correlation}
\usage{
polychoric_pi_set(
  pi,
  dist1,
  dist2,
  quant1,
  quant2,
  sd1,
  sd2,
  symmetric = FALSE
)
}
\arguments{
\item{pi}{Matrix of probabilities.}

\item{sd1, sd2}{Standard deviations of the distributions.}

\item{symmetric}{`TRUE` if the copula is symmetric, false otherwise.}

\item{dist_1, dist_2}{Distributions.}

\item{quant_1, quant_1}{Quantile functions of the distributions.}
}
\description{
Calculate the partial identification set for the latent correlation
}
