% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyiden.R
\name{polyserialiden}
\alias{polyserialiden}
\title{Calculate partial identification bounds for the polyserial correlation}
\usage{
polyserialiden(pi, marginals)
}
\arguments{
\item{pi}{Distribution function. Its first argument should be continous and its
second argument categorical.}

\item{marginals}{Specifies the marginal distribution. Either a named alternative
or a list containing the distribution functions, quantile functions, and
standard deviations for each marginal. Defaults to `"normal"`, which
uses standard normal marginals.}
}
\value{
The identification bounds.
}
\description{
https://www.rdocumentation.org/packages/polycor/versions/0.7-10/topics/polyserial
}
