% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reverse_pi}
\alias{reverse_pi}
\title{Reverse a pi matrix}
\usage{
reverse_pi(pi)
}
\arguments{
\item{pi}{A pi matrix}
}
\value{
The reversed matrix
}
\description{
The reverse pi matrix has the property that its maximal (minimal) correlation
   equals the negative of the minimal (maximal) correlation of the original pi
   matrix.
}
