% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{symmetrize}
\alias{symmetrize}
\title{Symmetrize a cumulative pi matrix}
\usage{
symmetrize(pi)
}
\arguments{
\item{pi}{A cumulative pi matrix.}
}
\value{
The symmetrized cumulative pi matrix.
}
\description{
Symmetrizes a cumulative pi matrix. Assumes the marginals have no
equal elements. Returns a symmetric matrix with \code{NA}s for elements that
are not known.
}
