% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{symmetry_constraints}
\alias{symmetry_constraints}
\title{Generate the constraints}
\usage{
symmetry_constraints(pi)
}
\arguments{
\item{pi}{The pi matrix.}
}
\value{
The symmetry constraints.
}
\description{
Generate the constraints
}
