% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetry.R
\name{symmetry_correlation}
\alias{symmetry_correlation}
\title{Find the upper and lower limits to the correlation assuming symmetry.}
\usage{
symmetry_correlation(
  pi,
  marginals = c("normal", "uniform", "exponential", "laplace"),
  method = c("substitution", "direct"),
  direction = c("upper", "lower"),
  max_rep = 1e+05
)
}
\arguments{
\item{pi}{The cumulative pi matrix.}

\item{marginals}{Specifies the marginal distribution. Either a named
alternative or a list containing the distribution functions, quantile
functions, and standard deviations for each marginal. (See the example.)
Defaults to \code{"normal"}, which uses standard normal marginals.}

\item{method}{The method used for integration. One of \code{"direct"} or
\code{"substitution"}. Defaults to \code{"substitution"}.}

\item{direction}{Direction of the correaltion.}

\item{max_rep}{Maximal repetitions when looking for an interior point.}
}
\value{
The upper / lower limit to the correlation.
}
\description{
Find the upper and lower limits to the correlation assuming symmetry.
}
