% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\encoding{UTF-8}
\name{transform_matrices}
\alias{transform_matrices}
\alias{pi_to_cum_pi}
\alias{cum_pi_to_pi}
\alias{reverse_pi}
\title{Transform probability matrices.}
\usage{
pi_to_cum_pi(pi)

cum_pi_to_pi(cum_pi)

reverse_pi(pi)
}
\arguments{
\item{pi, cum_pi}{The matrix of probabilities or matrix or cumulative probability matrix
of probabilities.}
}
\value{
The cumulative probability matrix corresponding to \code{pi} if \code{pi_to_cumpi} is
called; the probability matrix corresponding to \code{cum_pi} if \code{cum_pi_to_pi} is called. If
\code{reverse_pi} is called, it reverses the \code{pi} matrix.
}
\description{
The reverse pi matrix has the property that its maximal (minimal) correlation
equals the negative of the minimal (maximal) correlation of the original pi
matrix.
}
\examples{
   library("polyiden")
   copula = function(u) copula::pCopula(u, copula::normalCopula(0.3, dim = 2))
   pi = polyiden::generate_pi(i = 3, j = 3, copula = copula)
   polyiden::pi_to_cum_pi(pi)
   all(polyiden::cum_pi_to_pi((pi_to_cum_pi(pi))) == pi) # TRUE

   polyiden::polyiden(pi) # [1] -0.5699454  0.8549995
   -polyiden::polyiden(reverse_pi(pi)) # [1] 0.8549995 -0.5699454
}
\references{
Grønneberg, S., Moss, J., & Foldnes, N. (2020). Partial Identification of
Latent Correlations with Binary Data. Psychometrika, 85(4), 1028–1051.
https://doi.org/10.1007/s11336-020-09737-y
}
