% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyiden.R
\name{latent_correlation}
\alias{latent_correlation}
\title{Parametric latent correlations}
\usage{
latent_correlation(pi, copula, marginals, distribution)
}
\arguments{
\item{pi}{Matrix of probabilities.}

\item{copula}{Copula}

\item{marginals}{Specifies the marginal distribution. Either a named alternative
or a list containing the distribution functions, quantile functions, and
standard deviations for each marginal. Defaults to `"normal"`, which
uses standard normal marginals.}

\item{distribution}{Joint distribution. Ignored if `copula` is specified.}
}
\value{
The latent correlation.
}
\description{
Parametric latent correlations
}
