% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyiden.R
\encoding{UTF-8}
\name{latent_correlation}
\alias{latent_correlation}
\title{Parametric latent correlations}
\usage{
latent_correlation(
  pi,
  distribution = c("normal"),
  discrepancy = function(x, y) -y * log(x)
)
}
\arguments{
\item{pi}{Matrix of probabilities.}

\item{distribution}{The joint distribution of the latent variables,
parameterized by the correlation \code{"rho"}. Only \code{"normal"} supported.}

\item{discrepancy}{Discrepancy function to minimize. Defaults to
\verb{-y * log(x))}, which corresponds to minimization of the Kullback-Leibler
divergence.}
}
\value{
The closest latent correlation as judged by the discrepancy function.
}
\description{
Parametric latent correlations
}
\references{
Grønneberg, S., Moss, J., & Foldnes, N. (2020). Partial Identification of
Latent Correlations with Binary Data. Psychometrika, 85(4), 1028–1051.
https://doi.org/10.1007/s11336-020-09737-y
}
