% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_copulas.R
\name{limit_copula}
\alias{limit_copula}
\alias{lower_limit}
\alias{upper_limit}
\alias{lower_limit_cpp}
\alias{upper_limit_cpp}
\title{The limit copulas satisfying the constraints of pi}
\usage{
lower_limit(pi)

upper_limit(pi)

lower_limit_cpp(pi)

upper_limit_cpp(pi)
}
\arguments{
\item{pi}{Matrix of probabilities.}
}
\value{
The lower limit copula or upper limit copula.
}
\description{
The limit copulas satisfying the constraints of pi
}
