% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyiden.R
\name{polyserialiden}
\alias{polyserialiden}
\title{Calculate partial identification bounds for the polyserial correlation}
\usage{
polyserialiden(
  copula,
  points,
  marginals = c("normal", "uniform", "exponential", "laplace"),
  method = c("substitution", "direct")
)
}
\arguments{
\item{copula}{The copula.}

\item{points}{The points where the second argument is evaluated.}

\item{marginals}{Specifies the marginal distribution. Either a named alternative
or a list containing the distribution functions, quantile functions, and
standard deviations for each marginal. Defaults to \code{"normal"}, which
uses standard normal marginals.}

\item{method}{The method used for integration. One of \verb{direct"} or \verb{"substitution}. Defaults
to \verb{"substitution}.}
}
\value{
The identification bounds.
}
\description{
Calculate partial identification bounds for the polyserial correlation
}
